// src/interfaces/backend-bindings/ClientMessage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { JoinRoomMessage } from "./JoinRoomMessage";
import type { UpdateDocMessage } from "./UpdateDocMessage";

export type ClientMessage = { "type": "join_room", "data": JoinRoomMessage } | { "type": "update_doc", "data": UpdateDocMessage };