// src/interfaces/backend-bindings/ServerMessage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { UpdateDocMessage } from "./UpdateDocMessage";

export type ServerMessage = { "type": "confirm", "data": { message_type: string, message: string | null, } } | { "type": "update_doc", "data": UpdateDocMessage } | { "type": "error", "data": { message: string, } };